Removing type from badgePrimitive
