+++ src/Badge/_tests_/index.test.js
@@ -3,28 +3,30 @@
 import * as React from "react";
 import { shallow } from "enzyme";
 
-import NotificationBadge from "../index";
+import Badge from "../index";
 import Sightseeing from "../../icons/Sightseeing";
 
-describe("NotificationBadge", () => {
+describe("Badge", () => {
   const content = "badge";
   const type = "info";
   const dataTest = "test";
   const icon = <Sightseeing />;
   const ariaLabel = content;
 
-  const wrapped = shallow(
-    <NotificationBadge type={type} icon={icon} dataTest={dataTest} ariaLabel={ariaLabel}>
+  const component = shallow(
+    <Badge icon={icon} type={type} dataTest={dataTest} ariaLabel={ariaLabel}>
       {content}
-    </NotificationBadge>,
+    </Badge>,
   );
 
-  const component = wrapped.find("Badge");
-
   it("should have passed props", () => {
-    expect(component.prop("type")).toBe(type);
+    expect(component.prop("background")).toBe("#E5F7FF");
+    expect(component.prop("foregroundColor")).toBe("#0172CB");
     expect(component.render().prop("data-test")).toBe(dataTest);
     expect(component.render().prop("aria-label")).toBe(ariaLabel);
     expect(component.prop("icon")).toBe(icon);
+  });
+  it("should contain a content", () => {
+    expect(component.render().text()).toBe(content);
   });
 });
