+++ src/Badge/consts.js
