+++ src/primitives/BadgePrimitive/__tests__/index.test.js
@@ -8,7 +8,6 @@ import Sightseeing from "../../../icons/Sightseeing";
 
 describe("Badge", () => {
   const content = "badge";
-  const type = "info";
   const dataTest = "test";
   const icon = <Sightseeing />;
   const ariaLabel = content;
@@ -17,7 +16,6 @@ describe("Badge", () => {
 
   const component = shallow(
     <BadgePrimitive
-      type={type}
       icon={icon}
       background={background}
       foregroundColor={foregroundColor}
@@ -28,7 +26,6 @@ describe("Badge", () => {
     </BadgePrimitive>,
   );
   it("should have passed props", () => {
-    expect(component.prop("type")).toBe(type);
     expect(component.prop("background")).toBe(background);
     expect(component.prop("foregroundColor")).toBe(foregroundColor);
     expect(component.render().prop("data-test")).toBe(dataTest);
