+++ src/primitives/BadgePrimitive/index.js.flow
@@ -6,22 +6,8 @@ import { ReactComponentStyled } from "styled-components";
 
 import type { Globals } from "../../common/common.js.flow";
 
-export type Type =
-  | "neutral"
-  | "dark"
-  | "info"
-  | "success"
-  | "warning"
-  | "critical"
-  | "white"
-  | "infoInverted"
-  | "criticalInverted"
-  | "successInverted"
-  | "warningInverted";
-
 export type Props = {|
   +children?: React$Node,
-  +type?: Type,
   +icon?: React$Node,
   +ariaLabel?: string,
   +background?: string,
