+++ src/primitives/Primitives.stories.js
@@ -4,27 +4,26 @@ import { storiesOf } from "@storybook/react";
 import { text, select } from "@storybook/addon-knobs";
 
 import * as Icons from "../icons";
-import Badge from "./BadgePrimitive";
+import BadgePrimitive from "./BadgePrimitive";
 
 const getIcons = defaultIcon => select("Icon", [null, ...Object.keys(Icons)], defaultIcon);
 const getIcon = source => Icons[source];
 
 storiesOf("Primitives", module).add(
-  "Badge",
+  "BadgePrimitive",
   () => {
     const Icon = getIcon(getIcons("Airplane"));
-    const background = text("Background", "linear-gradient(red, yellow, green)");
-    const foregroundColor = text("ForegroundColor", "Blue");
+    const background = text("Background", "linear-gradient(#fd1d1d, #ffae28)");
+    const foregroundColor = text("ForegroundColor", "#fff");
 
     return (
-      <Badge
+      <BadgePrimitive
         background={background}
         foregroundColor={foregroundColor}
-        type="info"
         icon={Icon && <Icon />}
       >
-        Badge
-      </Badge>
+        BadgePrimitive
+      </BadgePrimitive>
     );
   },
   {
